Add CMakeLists.txr files into physx
The idea is to make integrating this project independent of python
and the pretty Nvidia specific `generate_project.py`.

The final goal is for other projects to FetchContent and add_subdirectory
or use ExternalProject in CMake and FindPhysX as any other standard
CMake project. But right now it only focus on building and installing
without needing python and with sane defaults with tweakable options.

The user just needs to run `CMake` in a `build-folder` not inside the
source tree (as usual for CMake projects), the top folder of the CMake
tree is `physx`.

This commit does not deal with the snippets or samples projects.

This commits only deals with new CMakeLists files added (heavily relying
on existing Nvidia CMake files).
The next commit modifies existing CMake files.
The split is just for clarity for external review.