[physx/source] Remove undefined behaviour from offset calclucations
The current code dereferences null pointers to compute an offset,
which has undefined behaviour. This patch changes that code to either
use the PX_OFFSET_OF macro (which is already used elsewhere for this
purpose, and which has well-defined behaviour on many popular
platforms), and creates a number of static member functions to expose
these values.